🚚 Restructure examples directory structure
To make the example implementations more modular, i.e., such that they can build on each other, their directory structure was entirely refactored.
