+++ examples/CMakeLists.txt
@@ -15,5 +15,13 @@
 #
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-add_subdirectory(device)
-add_subdirectory(driver)
+set(QDMI_EXAMPLES_INCLUDE_DIR "${CMAKE_CURRENT_SOURCE_DIR}/include")
+
+add_subdirectory(src)
+
+# add test code
+if(QDMI_BUILD_TESTS)
+  include(GoogleTest)
+  # include the tests itself
+  add_subdirectory(test)
+endif()
