--- examples/device/.clang-tidy
@@ -1,69 +1,0 @@
-FormatStyle: file
-
-Checks: |
-  bugprone-*,
-  -bugprone-easily-swappable-parameters,
-  clang-analyzer-*,
-  -clang-analyzer-security.insecureAPI.*,
-  clang-diagnostic-*,
-  cppcoreguidelines-*,
-  -cppcoreguidelines-non-private-member-variables-in-classes,
-  -cppcoreguidelines-avoid-magic-numbers,
-  -cppcoreguidelines-macro-usage,
-  -cppcoreguidelines-pro-bounds-constant-array-index,
-  -cppcoreguidelines-pro-bounds-array-to-pointer-decay,
-  -cppcoreguidelines-pro-bounds-pointer-arithmetic,
-  -cppcoreguidelines-owning-memory,
-  google-*,
-  -google-readability-todo,
-  -google-build-using-namespace,
-  misc-*,
-  -misc-no-recursion,
-  -misc-non-private-member-variables-in-classes,
-  modernize-*,
-  -modernize-use-trailing-return-type,
-  performance-*,
-  portability-*,
-  -portability-avoid-pragma-once
-  readability-*,
-  -readability-identifier-length,
-  -readability-magic-numbers,
-  -readability-function-cognitive-complexity
-
-CheckOptions:
-  - key: readability-identifier-naming.ClassCase
-    value: CamelCase
-  - key: readability-identifier-naming.ConstantParameterCase
-    value: lower_case
-  - key: readability-identifier-naming.EnumCase
-    value: UPPER_CASE
-  - key: readability-identifier-naming.EnumConstantCase
-    value: UPPER_CASE
-  - key: readability-identifier-naming.FunctionCase
-    value: Leading_upper_snake_case
-  - key: readability-identifier-naming.FunctionIgnoredRegexp
-    value: ".*QDMI.*"
-  - key: readability-identifier-naming.GlobalConstantCase
-    value: UPPER_CASE
-  - key: readability-identifier-naming.IgnoreMainLikeFunctions
-    value: "true"
-  - key: readability-identifier-naming.LocalConstantCase
-    value: lower_case
-  - key: readability-identifier-naming.MemberCase
-    value: lower_case
-  - key: readability-identifier-naming.MemberIgnoredRegexp
-    value: ".*QDMI.*"
-  - key: readability-identifier-naming.MethodCase
-    value: lower_case
-  - key: readability-identifier-naming.ParameterCase
-    value: lower_case
-  - key: readability-identifier-naming.NamespaceCase
-    value: lower_case
-  - key: readability-identifier-naming.StaticConstantCase
-    value: UPPER_CASE
-  - key: readability-identifier-naming.StructCase
-    value: Leading_upper_snake_case
-  - key: readability-identifier-naming.StructIgnoredRegexp
-    value: ".*QDMI.*"
-  - key: readability-identifier-naming.VariableCase
-    value: lower_case
