--- examples/device/.cmake-format.yaml
@@ -1,6 +1,0 @@
-format:
-  line_width: 80
-  keyword_case: "upper"
-  autosort: true
-markup:
-  first_comment_is_literal: true
