--- examples/device/.license-tools-config.json
@@ -1,41 +1,0 @@
-{
-  "author": {
-    "name": "QDMI Maintainers",
-    "years": [2024, 2026]
-  },
-  "force_author": true,
-  "custom_license": "Licensed under the Apache License v2.0 with LLVM Exceptions (the \"License\");\nyou may not use this file except in compliance with the License.\nYou may obtain a copy of the License at\n\nhttps://llvm.org/LICENSE.txt\n\nUnless required by applicable law or agreed to in writing, software\ndistributed under the License is distributed on an \"AS IS\" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the\nLicense for the specific language governing permissions and limitations under the License.\n\nSPDX-License-Identifier: Apache-2.0 WITH LLVM-exception",
-  "force_license": true,
-  "title": false,
-  "include": ["**/*"],
-  "style_override_for_suffix": {
-    ".pyi": "DOCSTRING_STYLE",
-    ".in": "POUND_STYLE",
-    ".mlir": "SLASH_STYLE",
-    ".td": "SLASH_STYLE",
-    ".yaml": "POUND_STYLE",
-    ".toml": "POUND_STYLE",
-    ".proto": "SLASH_STYLE",
-    ".ts": "C_STYLE",
-    ".css": "C_STYLE",
-    ".ps1": "POUND_STYLE"
-  },
-  "exclude": [
-    "^\\.[^/]+",
-    "/\\.[^/]+",
-    ".*\\.qasm",
-    ".*\\.ll",
-    ".*\\.md",
-    ".*\\.bib",
-    ".*\\.cff",
-    ".*\\.json",
-    ".*\\.html",
-    ".*\\.tfc",
-    ".*\\.qc",
-    ".*\\.real",
-    ".*\\.tex",
-    "uv\\.lock",
-    "py\\.typed",
-    "LICENSE"
-  ]
-}
