--- examples/device/cmake/ExternalDependencies.cmake
@@ -1,95 +1,0 @@
-# Copyright (c) 2024 - 2026 QDMI Maintainers
-# All rights reserved.
-#
-# Licensed under the Apache License v2.0 with LLVM Exceptions (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# https://llvm.org/LICENSE.txt
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-# License for the specific language governing permissions and limitations under
-# the License.
-#
-# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-include(FetchContent)
-include(CMakeDependentOption)
-set(FETCH_PACKAGES "")
-
-if(TARGET qdmi::qdmi)
-  message(STATUS "[qdmi][template] QDMI is already available.")
-else()
-  message(STATUS "[qdmi][template] QDMI will be included via FetchContent")
-  # cmake-format: off
-  set(QDMI_VERSION 1.3.2
-          CACHE STRING "QDMI version")
-  set(QDMI_REV "develop"
-          CACHE STRING "QDMI identifier (tag, branch or commit hash)")
-  set(QDMI_REPO_OWNER "Munich-Quantum-Software-Stack"
-          CACHE STRING "QDMI repository owner (change when using a fork)")
-  set(QDMI_INSTALL
-      OFF
-      CACHE BOOL "Generate installation instructions for QDMI")
-  # cmake-format: on
-  FetchContent_Declare(
-    qdmi
-    GIT_REPOSITORY https://github.com/${QDMI_REPO_OWNER}/qdmi.git
-    GIT_TAG ${QDMI_REV}
-    FIND_PACKAGE_ARGS ${QDMI_VERSION})
-  list(APPEND FETCH_PACKAGES qdmi)
-  message(
-    STATUS "[qdmi][template] QDMI will be included via FetchContent - done")
-endif()
-
-if(BUILD_CXX_QDMI_TESTS)
-  set(gtest_force_shared_crt
-      ON
-      CACHE BOOL "" FORCE)
-  set(GTEST_VERSION
-      1.17.0
-      CACHE STRING "Google Test version")
-  set(GTEST_URL
-      https://github.com/google/googletest/archive/refs/tags/v${GTEST_VERSION}.tar.gz
-  )
-  FetchContent_Declare(googletest URL ${GTEST_URL} FIND_PACKAGE_ARGS
-                                      ${GTEST_VERSION} NAMES GTest)
-  list(APPEND FETCH_PACKAGES googletest)
-endif()
-
-if(BUILD_CXX_QDMI_DOCS)
-  set(CMAKE_POLICY_DEFAULT_CMP0116
-      NEW
-      CACHE STRING
-            "Set the default CMP0116 policy to NEW for documentation builds")
-  set(MIN_DOXYGEN_VERSION
-      1.15.0
-      CACHE STRING "Minimum required Doxygen version")
-  set(DOXYGEN_REV
-      "669aeeefca743c148e2d935b3d3c69535c7491e6" # v1.16.1
-      CACHE STRING "Doxygen identifier (tag, branch or commit hash)")
-  FetchContent_Declare(
-    Doxygen
-    GIT_REPOSITORY https://github.com/doxygen/doxygen.git
-    GIT_TAG ${DOXYGEN_REV}
-    FIND_PACKAGE_ARGS ${MIN_DOXYGEN_VERSION})
-  list(APPEND FETCH_PACKAGES Doxygen)
-
-  set(DOXYGEN_AWESOME_VERSION
-      2.4.1
-      CACHE STRING "Doxygen Awesome version")
-  set(DOXYGEN_AWESOME_REV
-      "1f3620084ff75734ed192101acf40e9dff01d848"
-      CACHE STRING "Doxygen Awesome identifier (tag, branch or commit hash)")
-  FetchContent_Declare(
-    doxygen-awesome-css
-    GIT_REPOSITORY https://github.com/jothepro/doxygen-awesome-css.git
-    GIT_TAG ${DOXYGEN_AWESOME_REV}
-    FIND_PACKAGE_ARGS ${DOXYGEN_AWESOME_VERSION})
-  list(APPEND FETCH_PACKAGES doxygen-awesome-css)
-endif()
-
-# Make all declared dependencies available.
-FetchContent_MakeAvailable(${FETCH_PACKAGES})
