--- examples/device/cmake/cxx-qdmi-device-config.cmake.in
@@ -1,32 +1,0 @@
-# Copyright (c) 2024 - 2026 QDMI Maintainers
-# All rights reserved.
-#
-# Licensed under the Apache License v2.0 with LLVM Exceptions (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# https://llvm.org/LICENSE.txt
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-# License for the specific language governing permissions and limitations under
-# the License.
-#
-# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-# A CMake config file for the library, to be used by external projects
-
-@PACKAGE_INIT@
-
-list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_LIST_DIR}")
-
-if(TARGET cxx-qdmi-device)
-  return()
-endif()
-
-include("${CMAKE_CURRENT_LIST_DIR}/cxx-qdmi-device-targets.cmake")
-
-if(NOT cxx-qdmi-device_FIND_QUIETLY)
-  message(STATUS "Found CXX QDMI Device version ${cxx-qdmi-device_VERSION}")
-endif()
