--- examples/device/docs/header.html
@@ -1,196 +1,0 @@
-<!-- HTML header generated by doxygen 1.12.0 via `doxygen -l layout.xml` -->
-<!-- Included by the Doxygen.in file under the tag `HTML_HEADER` -->
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "https://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml" lang="$langISO">
-  <head>
-    <meta http-equiv="Content-Type" content="text/xhtml;charset=UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=11" />
-    <meta name="generator" content="Doxygen $doxygenversion" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <!--BEGIN PROJECT_NAME-->
-    <title>$projectname: $title</title>
-    <!--END PROJECT_NAME-->
-    <!--BEGIN !PROJECT_NAME-->
-    <title>$title</title>
-    <!--END !PROJECT_NAME-->
-    <!--BEGIN PROJECT_ICON-->
-    <link rel="icon" href="$relpath^$projecticon" type="image/x-icon" />
-    <!--END PROJECT_ICON-->
-    <link href="$relpath^tabs.css" rel="stylesheet" type="text/css" />
-    <!--BEGIN DISABLE_INDEX-->
-    <!--BEGIN FULL_SIDEBAR-->
-    <script type="text/javascript">
-      var page_layout = 1;
-    </script>
-    <!--END FULL_SIDEBAR-->
-    <!--END DISABLE_INDEX-->
-    <script type="text/javascript" src="$relpath^jquery.js"></script>
-    <script type="text/javascript" src="$relpath^dynsections.js"></script>
-    <!--BEGIN COPY_CLIPBOARD-->
-    <script type="text/javascript" src="$relpath^clipboard.js"></script>
-    <!--END COPY_CLIPBOARD-->
-    <script
-      type="text/javascript"
-      src="$relpath^doxygen-awesome-darkmode-toggle.js"
-    ></script>
-    <script
-      type="text/javascript"
-      src="$relpath^doxygen-awesome-fragment-copy-button.js"
-    ></script>
-    <script
-      type="text/javascript"
-      src="$relpath^doxygen-awesome-paragraph-link.js"
-    ></script>
-    <script
-      type="text/javascript"
-      src="$relpath^doxygen-awesome-interactive-toc.js"
-    ></script>
-    <script
-      type="text/javascript"
-      src="$relpath^doxygen-awesome-tabs.js"
-    ></script>
-    <script type="text/javascript">
-      DoxygenAwesomeDarkModeToggle.init();
-      DoxygenAwesomeFragmentCopyButton.init();
-      DoxygenAwesomeParagraphLink.init();
-      DoxygenAwesomeInteractiveToc.init();
-      DoxygenAwesomeTabs.init();
-    </script>
-    $treeview $search $mathjax $darkmode
-    <link href="$relpath^$stylesheet" rel="stylesheet" type="text/css" />
-    $extrastylesheet
-  </head>
-  <body>
-    <!-- https://tholman.com/github-corners/ -->
-    <a
-      href="https://github.com/Munich-Quantum-Software-Stack/QDMI"
-      class="github-corner"
-      title="View source on GitHub"
-      target="_blank"
-      rel="noopener noreferrer"
-    >
-      <svg
-        viewBox="0 0 250 250"
-        width="40"
-        height="40"
-        style="position: absolute; top: 0; border: 0; right: 0; z-index: 99"
-        aria-hidden="true"
-      >
-        <path d="M0,0 L115,115 L130,115 L142,142 L250,250 L250,0 Z"></path>
-        <path
-          d="M128.3,109.0 C113.8,99.7 119.0,89.6 119.0,89.6 C122.0,82.7 120.5,78.6 120.5,78.6 C119.2,72.0 123.4,76.3 123.4,76.3 C127.3,80.9 125.5,87.3 125.5,87.3 C122.9,97.6 130.6,101.9 134.4,103.2"
-          fill="currentColor"
-          style="transform-origin: 130px 106px"
-          class="octo-arm"
-        ></path>
-        <path
-          d="M115.0,115.0 C114.9,115.1 118.7,116.5 119.8,115.4 L133.7,101.6 C136.9,99.2 139.9,98.4 142.2,98.6 C133.8,88.0 127.5,74.4 143.8,58.0 C148.5,53.4 154.0,51.2 159.7,51.0 C160.3,49.4 163.2,43.6 171.4,40.1 C171.4,40.1 176.1,42.5 178.8,56.2 C183.1,58.6 187.2,61.8 190.9,65.4 C194.5,69.0 197.7,73.2 200.1,77.6 C213.8,80.2 216.3,84.9 216.3,84.9 C212.7,93.1 206.9,96.0 205.4,96.6 C205.1,102.4 203.0,107.8 198.3,112.5 C181.9,128.9 168.3,122.5 157.7,114.1 C157.9,116.9 156.7,120.9 152.7,124.9 L141.0,136.5 C139.8,137.7 141.6,141.9 141.8,141.8 Z"
-          fill="currentColor"
-          class="octo-body"
-        ></path>
-      </svg>
-    </a>
-    <style>
-      .github-corner:hover .octo-arm {
-        animation: octocat-wave 560ms ease-in-out;
-      }
-
-      @keyframes octocat-wave {
-        0%,
-        100% {
-          transform: rotate(0);
-        }
-        20%,
-        60% {
-          transform: rotate(-25deg);
-        }
-        40%,
-        80% {
-          transform: rotate(10deg);
-        }
-      }
-
-      @media (max-width: 500px) {
-        .github-corner:hover .octo-arm {
-          animation: none;
-        }
-
-        .github-corner .octo-arm {
-          animation: octocat-wave 560ms ease-in-out;
-        }
-      }
-
-      /* image path depending on the color scheme */
-      /* Default (light mode) */
-      .logo {
-        width: 130px;
-        height: auto;
-        display: block;
-      }
-      .logo {
-        content: url("$relpath^$projectlogo"); /* this is set by doxygen, see PROJECT_LOGO in the Doxyfile.in */
-      }
-      /* Dark mode */
-      html.dark-mode .logo {
-        content: url("$relpath^cxx_logo_dark.svg");
-      }
-    </style>
-    <!--BEGIN DISABLE_INDEX-->
-    <!--BEGIN FULL_SIDEBAR-->
-    <div id="side-nav" class="ui-resizable side-nav-resizable">
-      <!-- do not remove this div, it is closed by doxygen! -->
-      <!--END FULL_SIDEBAR-->
-      <!--END DISABLE_INDEX-->
-
-      <div id="top">
-        <!-- do not remove this div, it is closed by doxygen! -->
-
-        <!--BEGIN TITLEAREA-->
-        <div id="titlearea">
-          <table cellspacing="0" cellpadding="0">
-            <tbody>
-              <tr id="projectrow">
-                <!--BEGIN PROJECT_LOGO-->
-                <td id="projectlogo"><img class="logo" alt="Logo" /></td>
-                <!--END PROJECT_LOGO-->
-                <!--BEGIN PROJECT_NAME-->
-                <td id="projectalign">
-                  <div id="projectname">
-                    $projectname<!--BEGIN PROJECT_NUMBER--><span
-                      id="projectnumber"
-                      >&#160;$projectnumber</span
-                    >
-                    <!--END PROJECT_NUMBER-->
-                  </div>
-                  <!--BEGIN PROJECT_BRIEF-->
-                  <div id="projectbrief">$projectbrief</div>
-                  <!--END PROJECT_BRIEF-->
-                </td>
-                <!--END PROJECT_NAME-->
-                <!--BEGIN !PROJECT_NAME-->
-                <!--BEGIN PROJECT_BRIEF-->
-                <td>
-                  <div id="projectbrief">$projectbrief</div>
-                </td>
-                <!--END PROJECT_BRIEF-->
-                <!--END !PROJECT_NAME-->
-                <!--BEGIN DISABLE_INDEX-->
-                <!--BEGIN SEARCHENGINE-->
-                <!--BEGIN !FULL_SIDEBAR-->
-                <td>$searchbox</td>
-                <!--END !FULL_SIDEBAR-->
-                <!--END SEARCHENGINE-->
-                <!--END DISABLE_INDEX-->
-              </tr>
-              <!--BEGIN SEARCHENGINE-->
-              <!--BEGIN FULL_SIDEBAR-->
-              <tr>
-                <td colspan="2">$searchbox</td>
-              </tr>
-              <!--END FULL_SIDEBAR-->
-              <!--END SEARCHENGINE-->
-            </tbody>
-          </table>
-        </div>
-        <!--END TITLEAREA-->
-        <!-- end header part -->
