--- examples/device/docs/style.css
@@ -1,50 +1,0 @@
-/*
- * Copyright (c) 2024 - 2026 QDMI Maintainers
- * All rights reserved.
- *
- * Licensed under the Apache License v2.0 with LLVM Exceptions (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://llvm.org/LICENSE.txt
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- *
- * SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
- */
-
-/* Generated by doxygen 1.12.0 via `doxygen -w header.html footer.html style.css` */
-/* Included in the Doxygen.in file under the tag HTML_EXTRA_STYLESHEETS */
-
-/* taken from awesome-doxygen-css to scale the github link in the upper right corner */
-.github-corner svg {
-  fill: var(--primary-light-color);
-  color: var(--page-background-color);
-  width: 72px;
-  height: 72px;
-}
-
-@media screen and (max-width: 767px) {
-  .github-corner svg {
-    width: 50px;
-    height: 50px;
-  }
-
-  #projectnumber {
-    margin-right: 22px;
-  }
-}
-
-/* add some space between version number and project title */
-#projectnumber {
-  margin-left: 10px;
-}
-
-/* increase size of the logo area to fit the two line brief description */
-html {
-  --top-height: 160px;
-}
