--- examples/device/python/cxx/qdmi/py.typed
@@ -1,2 +1,0 @@
-# Instruct type checkers to look for inline type annotations in this package.
-# See PEP 561.
