+++ examples/src/CMakeLists.txt
@@ -15,18 +15,5 @@
 #
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-"""Site customization shim to enable multiprocess coverage collection in tests.
-
-See https://coverage.readthedocs.io/en/latest/subprocess.html.
-"""
-
-from __future__ import annotations
-
-try:
-    import coverage
-
-    coverage.process_startup()
-except ImportError:
-    # The 'coverage' module is optional
-    # If it is not installed, we do not enable multiprocess coverage collection
-    pass
+add_subdirectory(device)
+add_subdirectory(driver)
