+++ examples/src/device/cxx_device.cpp
