+++ examples/test/CMakeLists.txt
@@ -15,15 +15,4 @@
 #
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-# add C++ language support
-enable_language(CXX)
-
-add_library(qdmi_example_driver qdmi_example_driver.cpp qdmi_example_driver.h)
-target_link_libraries(qdmi_example_driver PRIVATE qdmi::qdmi
-                                                  qdmi::qdmi_project_warnings)
-target_include_directories(qdmi_example_driver
-                           PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
-target_compile_features(qdmi_example_driver PRIVATE cxx_std_20)
-set_target_properties(qdmi_example_driver PROPERTIES POSITION_INDEPENDENT_CODE
-                                                     ON)
-add_library(qdmi::example_driver ALIAS qdmi_example_driver)
+add_subdirectory(device)
